# sdk-server/docs/swagger.yaml
definitions:
  base.PageResp:
    properties:
      currentPage:
        description: 当前第几页
        type: integer
      list:
        description: 数据列表
      pageSize:
        description: 分页大小
        type: integer
      total:
        description: 总条数
        type: integer
    type: object
  base.ReqId:
    properties:
      id:
        description: 主键ID
        type: integer
    type: object
  base.ReqIds:
    properties:
      ids:
        description: 多id
        items:
          type: integer
        type: array
    type: object
  base.Resp:
    properties:
      code:
        description: 返回码
        type: integer
      data:
        description: 数据
      msg:
        description: 消息
        type: string
      reqId:
        description: '`json:"请求id"`'
        type: string
    type: object
...
  dto.AdminDto:
    properties:
      avatar:
        description: 头像
        type: string
      bio:
        description: 签名
        type: string
      birthday:
        description: 生日 格式 yyyy-MM-dd
        type: string
      clientId:
        description: 客户端id
        type: string
      deptId:
        description: 部门id
        type: integer
      email:
        description: 邮箱
        type: string
      gender:
        description: 性别 1男 2女 3未知
        type: integer
      id:
        description: 主键
        type: integer
      ipLocation:
        type: string
      lockTime:
        description: 锁定结束时间
        type: string
      name:
        description: 姓名
        type: string
      nickname:
        description: 昵称
        type: string
      password:
        description: 密码
        type: string
      phone:
        description: 手机号
        type: string
      regIp:
        description: 注册ip
        type: string
      remark:
        description: 备注
        type: string
      roleId:
        description: 角色id
        type: integer
      status:
        description: 状态 1正常
        type: integer
      username:
        description: 用户名
        type: string
    type: object
  dto.AdminGetPageReq:
    properties:
      deptId:
        description: 部门id
        type: integer
      name:
        description: 姓名
        type: string
      page:
        description: 页码
        type: integer
      pageSize:
        description: 每页大小
        type: integer
      phone:
        description: 手机号
        type: string
      status:
        description: 状态 1正常
        type: integer
    type: object
...
  dto.CounterDto:
    properties:
      key:
        description: 主键
        type: string
      val:
        description: 值
        type: integer
    type: object
  dto.CounterGetPageReq:
    properties:
      page:
        description: 页码
        type: integer
      pageSize:
        description: 每页大小
        type: integer
    type: object
...
  dto.NoticeDto:
    properties:
      count:
        type: integer
      key:
        type: string
      list:
        items:
          $ref: '#/definitions/dto.NoticeItem'
        type: array
      name:
        type: string
      total:
        type: integer
    type: object
  dto.NoticeItem:
    properties:
      beginAt:
        type: integer
      content:
        type: string
      createdAt:
        type: integer
      endAt:
        type: integer
      id:
        type: integer
      noticeType:
        type: integer
      op:
        type: integer
      opId:
        type: integer
      reminderStatus:
        type: integer
      reminderTime:
        type: integer
      status:
        type: integer
      title:
        type: string
      type:
        type: integer
    type: object
...
  dto.TaskDto:
    properties:
      beginAt:
        description: 开始时间
        type: string
      content:
        description: 任务内容
        type: string
      endAt:
        description: 结束时间
        type: string
      id:
        description: 主键
        type: integer
      op:
        description: 操作类型
        type: integer
      opId:
        description: 操作id
        type: integer
      reminderStatus:
        description: 提醒状态 1开启 2关闭
        type: integer
      reminderTime:
        description: 提醒时间
        type: string
      status:
        description: 状态1开启2关闭
        type: integer
      taskType:
        description: 任务类型
        type: integer
      teamId:
        description: 团队id
        type: integer
      title:
        description: 任务标题
        type: string
      userId:
        description: 用户id
        type: integer
    type: object
  dto.TaskGetPageReq:
    properties:
      page:
        description: 页码
        type: integer
      pageSize:
        description: 每页大小
        type: integer
      status:
        description: 状态1开启2关闭ReminderStatus int `json:"reminderStatus" query:"column:reminder_status"`
          //提醒状态 1开启 2关闭
        type: integer
      teamId:
        type: integer
      userId:
        type: integer
    type: object
...
  models.Admin:
    properties:
      avatar:
        description: 头像
        type: string
      bio:
        description: 签名
        type: string
      birthday:
        description: 生日 格式 yyyy-MM-dd
        type: string
      clientId:
        description: 客户端id
        type: string
      createdAt:
        description: 创建时间
        type: string
      deptId:
        description: 部门id
        type: integer
      deptPath:
        description: 部门路径
        type: string
      email:
        description: 邮箱
        type: string
      gender:
        description: 性别 1男 2女 3未知
        type: integer
      id:
        description: 主键
        type: integer
      ipLocation:
        type: string
      lockTime:
        description: 锁定结束时间
        type: string
      name:
        description: 姓名
        type: string
      nickname:
        description: 昵称
        type: string
      password:
        description: 密码
        type: string
      phone:
        description: 手机号
        type: string
      regIp:
        description: 注册ip
        type: string
      remark:
        description: 备注
        type: string
      roleId:
        description: 角色id
        type: integer
      status:
        description: 状态 1正常
        type: integer
      updateBy:
        description: 更新者
        type: integer
      updatedAt:
        description: 最后更新时间
        type: string
      username:
        description: 用户名
        type: string
    type: object
...
  models.Counter:
    properties:
      key:
        description: 主键
        type: string
      seq:
        description: 值
        type: integer
    type: object
...
  models.Task:
    properties:
      beginAt:
        description: 开始时间
        type: string
      content:
        description: 任务内容
        type: string
      createdAt:
        description: 创建时间
        type: string
      endAt:
        description: 结束时间
        type: string
      id:
        description: 主键
        type: integer
      op:
        description: 操作类型
        type: integer
      opId:
        description: 操作id
        type: integer
      reminderStatus:
        description: 提醒状态 1开启 2关闭
        type: integer
      reminderTime:
        description: 提醒时间
        type: string
      status:
        description: 状态1开启2关闭
        type: integer
      taskType:
        description: 任务类型
        type: integer
      teamId:
        description: 团队id
        type: integer
      title:
        description: 任务标题
        type: string
      updatedAt:
        description: 更新时间
        type: string
      userId:
        description: 用户id
        type: integer
    type: object
...
info:
  contact:
    email: baowk1024@gmail.com
    name: baowk
    url: https://github.com/baowk/dilu
  description: 致力于做一个开发快速，运行稳定的框架
  title: Dilu API
  version: V0.0.1
paths:
...
  /api/notice/task/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.TaskDto'
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/base.Resp'
            - properties:
                data:
                  $ref: '#/definitions/models.Task'
              type: object
      security:
      - Bearer: []
      summary: 创建Task
      tags:
      - notice-Task
  /api/notice/task/del:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/base.ReqIds'
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/base.Resp'
            - properties:
                data:
                  $ref: '#/definitions/models.Task'
              type: object
      security:
      - Bearer: []
      summary: 删除Task
      tags:
      - notice-Task
  /api/notice/task/get:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/base.ReqId'
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/base.Resp'
            - properties:
                data:
                  $ref: '#/definitions/models.Task'
              type: object
      security:
      - Bearer: []
      summary: 获取Task
      tags:
      - notice-Task
  /api/notice/task/my:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.TaskGetPageReq'
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/base.Resp'
            - properties:
                data:
                  $ref: '#/definitions/dto.NoticeDto'
              type: object
      security:
      - Bearer: []
      summary: 获取Task列表
      tags:
      - notice-Task
  /api/notice/task/page:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.TaskGetPageReq'
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/base.Resp'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/base.PageResp'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/models.Task'
                        type: array
                    type: object
              type: object
      security:
      - Bearer: []
      summary: 获取Task列表
      tags:
      - notice-Task
  /api/notice/task/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.TaskDto'
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/base.Resp'
            - properties:
                data:
                  $ref: '#/definitions/models.Task'
              type: object
      security:
      - Bearer: []
      summary: 更新Task
      tags:
      - notice-Task
...
  /api/sys/admin/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.AdminDto'
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/base.Resp'
            - properties:
                data:
                  $ref: '#/definitions/models.Admin'
              type: object
      security:
      - Bearer: []
      summary: 创建用户
      tags:
      - sys-Admin
  /api/sys/admin/del:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/base.ReqIds'
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/base.Resp'
            - properties:
                data:
                  $ref: '#/definitions/models.Admin'
              type: object
      security:
      - Bearer: []
      summary: 删除用户
      tags:
      - sys-Admin
...
  /api/sys/admin/get:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/base.ReqId'
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/base.Resp'
            - properties:
                data:
                  $ref: '#/definitions/models.Admin'
              type: object
      security:
      - Bearer: []
      summary: 获取用户
      tags:
      - sys-Admin
...
  /api/sys/admin/page:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.AdminGetPageReq'
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/base.Resp'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/base.PageResp'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/models.Admin'
                        type: array
                    type: object
              type: object
      security:
      - Bearer: []
      summary: 获取用户列表
      tags:
      - sys-Admin
...
  /api/sys/admin/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.AdminDto'
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/base.Resp'
            - properties:
                data:
                  $ref: '#/definitions/models.Admin'
              type: object
      security:
      - Bearer: []
      summary: 更新用户
      tags:
      - sys-Admin
...
  /api/sys/counter/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.CounterDto'
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/base.Resp'
            - properties:
                data:
                  $ref: '#/definitions/models.Counter'
              type: object
      security:
      - Bearer: []
      summary: 创建计数器
      tags:
      - sys-Counter
  /api/sys/counter/del:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/base.ReqIds'
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/base.Resp'
            - properties:
                data:
                  $ref: '#/definitions/models.Counter'
              type: object
      security:
      - Bearer: []
      summary: 删除计数器
      tags:
      - sys-Counter
  /api/sys/counter/get:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/base.ReqId'
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/base.Resp'
            - properties:
                data:
                  $ref: '#/definitions/models.Counter'
              type: object
      security:
      - Bearer: []
      summary: 获取计数器
      tags:
      - sys-Counter
  /api/sys/counter/page:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.CounterGetPageReq'
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/base.Resp'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/base.PageResp'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/models.Counter'
                        type: array
                    type: object
              type: object
      security:
      - Bearer: []
      summary: 获取计数器列表
      tags:
      - sys-Counter
  /api/sys/counter/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.CounterDto'
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/base.Resp'
            - properties:
                data:
                  $ref: '#/definitions/models.Counter'
              type: object
      security:
      - Bearer: []
      summary: 更新计数器
      tags:
      - sys-Counter
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"